[FasterRCNN] always save the last model

+++ tensorpack/callbacks/saver.py
@@ -25,7 +25,7 @@ class ModelSaver(Callback):
         """
         Args:
             max_to_keep (int): the same as in ``tf.train.Saver``.
-            keep_checkpoint_every_n_hours (int): the same as in ``tf.train.Saver``.
+            keep_checkpoint_every_n_hours (float): the same as in ``tf.train.Saver``.
             checkpoint_dir (str): Defaults to ``logger.get_logger_dir()``.
             var_collections (str or list of str): collection of the variables (or list of collections) to save.
         """
